pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dns-probe-system"
version = "0.1.0"
description = "DNS probe system MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.43",
    "pydantic>=2.11.7",
    "jinja2>=3.1.6",
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "psycopg[binary]",
    "pymysql>=1.1.0",
    # MySQL 8 默认 caching_sha2_password，PyMySQL 需 cryptography 做 RSA
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["apps*", "packages*"]

[tool.setuptools.package-data]
"apps.api" = ["static/**/*"]

[tool.uvicorn]
factory = false
reload = true
